+++ myboot-demo/myboot-admin-web/src/main/java/org/siu/myboot/server/entity/po/QProductBrand.java
@@ -20,7 +20,7 @@ public class QProductBrand extends EntityPathBase<ProductBrand> implements QBuil
 
 	public static final QProductBrand productBrand = new QProductBrand("productBrand");
 
-	public final NumberPath<Integer> id = createNumber("id", Integer.class);
+	public final NumberPath<Long> id = createNumber("id", Long.class);
 
 	public final StringPath brandName = createString("brandName");
