+++ myboot-demo/myboot-admin-web/src/main/java/org/siu/myboot/server/entity/po/QProductSupplier.java
@@ -20,7 +20,7 @@ public class QProductSupplier extends EntityPathBase<ProductSupplier> implements
 
 	public static final QProductSupplier productSupplier = new QProductSupplier("productSupplier");
 
-	public final NumberPath<Integer> id = createNumber("id", Integer.class);
+	public final NumberPath<Long> id = createNumber("id", Long.class);
 
 	public final StringPath supplierName = createString("supplierName");
